requirements.txt
langchain>=0.0.267
openai>=1.3.0
chromadb>=0.4.13
tiktoken>=0.4.0
python-dotenv>=1.0.0
pyyaml>=6.0
numpy>=1.24.3
SpeechRecognition>=3.10.0
pydantic>=2.3.0
pyaudio>=0.2.13
elevenlabs>=0.2.24
transformers>=4.33.1
torch>=2.0.1
fastapi>=0.103.1
uvicorn>=0.23.2
python-telegram-bot>=20.5
discord.py>=2.3.2
twilio>=8.5.0  # For WhatsApp integration
rich>=13.5.2